src/maze_resnet/requirements.txt
# Python 3.10.12
numpy==1.23.5
numpy==1.21.5
Pillow==9.0.1
Pillow==11.1.0
torch==2.5.1+cu121
torchvision==0.20.1+cu121
